now drawing lines instead of points
Fixed issue where fast moving objects/particles were drawn as single
points on the canvas. Now objects/particles also store a previous
position in order to draw a line between the previous position and the
new position, filling the space between the two position which was not
drawn between frames.