improved mobile support
now uses width detection to detect if a mobile device is used.
If a mobile device is used: the search should (hopefully) not be able to
show up.
Now if useScene is False the app automatically sets itself on a
non-existing scene, displaying the background.
fixed box-shadow bug for clock and list